+++ v2/Optimize/Optimize/GlyphData.h
@@ -15,7 +15,7 @@
 - (void)dealloc;
 @property (readonly) NSArray<NSDictionary<NSString *, id> *> *urlData;
 @property (readonly) NSArray<NSNumber *> *glyphSizes;
-@property (readonly) uint32_t* urlBitmaps;
+@property (readonly) uint8_t* urlBitmaps;
 @property (readonly) NSUInteger glyphCount;
 @property (readonly) NSUInteger glyphBitfieldSize;
 @property (readonly) NSUInteger urlCount;
