+++ v2/Optimize/Optimize/PickAndPlace.metal
@@ -56,7 +56,7 @@ private:
     int32_t offset{0};
 };
 
-kernel void possibleFitnesses(device uint32_t* generation [[buffer(0)]], device uint32_t* glyphSizes [[buffer(1)]], device uint32_t* urlBitmaps [[buffer(2)]], device uint32_t* output [[buffer(3)]], constant uint32_t& glyphIndex [[buffer(4)]], uint3 tid [[thread_position_in_grid]]) {
+kernel void possibleFitnesses(device uint32_t* generation [[buffer(0)]], device uint32_t* glyphSizes [[buffer(1)]], device uint8_t* urlBitmaps [[buffer(2)]], device uint32_t* output [[buffer(3)]], constant uint32_t& glyphIndex [[buffer(4)]], uint3 tid [[thread_position_in_grid]]) {
     uint generationIndex = tid.x;
     uint urlIndex = tid.y;
     uint rotationIndex = tid.z;
